Moved template redirect definitions to separate file for ease of access
